+++ src/cone/app/workflow.py
@@ -32,6 +32,15 @@ def persist_state(node, info):
     node()
 
 
+def permission_checker(permission, node, request):
+    """Callback for repoze.workflow to check permissions.
+
+    Necessary since pyramid 1.8, where ``pyramid.security.has_permission``
+    has been removed
+    """
+    return request.has_permission(permission, node)
+
+
 @implementer(IWorkflowState)
 class WorkflowState(Behavior):
     """Behavior for nodes providing workflow states.